renaming model to specfem_model to make it clear that this is a specific class basic functionality for model input and serial looping created bombed out a few functionalities and merged a few functions together
